backend/requirements.txt
pytesseract
python-pptx
Pillow
python-dotenv
pymupdf
markitdown
easyocr
torch
torchvision
opencv-python-headless
chromadb
sentence-transformers
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
accelerate
